refactor(quant): 重构量化交易系统核心组件
- 修改手续费计算接口将成交量参数从long改为int类型
- 将持仓策略接口从enter/exit方法改为calcEnterVolume/calcExitVolume方法
- 移除TradeLog类中的构建器注解并调整构造函数参数顺序
- 重新设计Trader为抽象接口并创建SimulationTrader实现类
- 新增AbstractTrader作为交易者的基础实现类
- 更新测试代码适配新的交易者初始化方式

+++ clever-task-example/src/main/java/org/clever/quant/TradeFeeStrategy.java
@@ -13,7 +13,7 @@ public interface TradeFeeStrategy {
      * @param price  撮合成交价
      * @param volume 成交量
      */
-    double calcEnterFee(double price, long volume);
+    double calcEnterFee(double price, int volume);
 
     /**
      * 计算平仓手续费
@@ -21,5 +21,5 @@ public interface TradeFeeStrategy {
      * @param price  撮合成交价
      * @param volume 成交量
      */
-    double calcExitFee(double price, long volume);
+    double calcExitFee(double price, int volume);
 }
